+++ src/Table.php
@@ -400,6 +400,35 @@ class Table implements ArrayAccess, Countable
         return $this->fields;
     }
 
+    public function __call(string $name, array $args): ?Row
+    {
+        $field = $this->__get($name);
+
+        return $this->select()->one()->where("{$field} = ", $args[0])->run();
+    }
+
+    /**
+     * Search a row with some values or create one if it does not exist
+     */
+    public function getOrCreate(array $data): Row
+    {
+        $query = $this->select()->one();
+
+        foreach ($data as $name => $value) {
+            $field = $this->__get($name);
+
+            $query->where("{$field} = ", $value);
+        }
+
+        $row = $query->run();
+
+        if ($row) {
+            return $row;
+        }
+
+        return $this->create($data);
+    }
+
     public function create(array $data = []): Row
     {
         if (isset($data['id']) && ($row = $this->getCached($data['id']))) {